# pkg/api/rest/swaggerDocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CB-Myfw REST API",
        "title": "CB-Myfw REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://cloud-barista.github.io",
            "email": "contact-to-cloud-barista@googlegroups.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.3.0-espresso"
    },
    "host": "localhost:4321",
    "basePath": "/myfw",
    "paths": {
        "/healthcheck": {
            "get": {
                "description": "for health check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Default"
                ],
                "summary": "Health Check",
                "operationId": "HealthCheck",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ns/{namespace}/clusters": {
            "get": {
                "description": "List Cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "List Cluster",
                "operationId": "ListCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ClusterList"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "Create Cluster",
                "operationId": "CreateCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reuest json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ClusterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Cluster"
                        }
                    }
                }
            }
        },
        "/ns/{namespace}/clusters/{cluster}": {
            "get": {
                "description": "Get Cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "Get Cluster",
                "operationId": "GetCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Cluster"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "Delete a cluster",
                "operationId": "DeleteCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Status"
                        }
                    }
                }
            }
        },
        "/ns/{namespace}/clusters/{cluster}/nodes": {
            "get": {
                "description": "List Node",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "List Node",
                "operationId": "ListNode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.NodeList"
                        }
                    }
                }
            },
            "post": {
                "description": "Add Node",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "Add Node",
                "operationId": "AddNode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reuest json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NodeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Node"
                        }
                    }
                }
            }
        },
        "/ns/{namespace}/clusters/{cluster}/nodes/{node}": {
            "get": {
                "description": "Get Node",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "Get Node",
                "operationId": "GetNode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "node",
                        "name": "node",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Node"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove Node",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "Remove Node",
                "operationId": "RemoveNode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "node",
                        "name": "node",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Status"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Cluster": {
            "type": "object",
            "properties": {
                "clusterConfig": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "mcis": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Node"
                    }
                },
                "status": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "model.ClusterList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Cluster"
                    }
                },
                "kind": {
                    "type": "string"
                }
            }
        },
        "model.ClusterReq": {
            "type": "object",
            "properties": {
                "controlPlaneNodeCount": {
                    "type": "integer"
                },
                "controlPlaneNodeSpec": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "workerNodeCount": {
                    "type": "integer"
                },
                "workerNodeSpec": {
                    "type": "string"
                }
            }
        },
        "model.Node": {
            "type": "object",
            "properties": {
                "credential": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "publicIp": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "spec": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "model.NodeList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Node"
                    }
                },
                "kind": {
                    "type": "string"
                }
            }
        },
        "model.NodeReq": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "string"
                },
                "workerNodeCount": {
                    "type": "integer"
                },
                "workerNodeSpec": {
                    "type": "string"
                }
            }
        },
        "model.Status": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}